pkp/pkp-lib#2072 Fix unit tests
